updated sample reception component

+++ app/Http/Livewire/Lab/SampleManagement/SpecimenRequestComponent.php
@@ -746,15 +746,29 @@ class SpecimenRequestComponent extends Component
 
     public function deleteData()
     {
+        $sample = Sample::where('id', $this->delete_id)->first();
         try {
-            $sample = Sample::where('id', $this->delete_id)->first();
-            $sample->delete();
-            $this->delete_id = '';
-            $this->dispatchBrowserEvent('close-modal');
-            $this->dispatchBrowserEvent('alert', ['type' => 'success',  'message' => 'Sample Information deleted successfully!']);
+            if($sample->status=='Accessioned'){
+                $sample->delete();
+                $this->delete_id = '';
+                $this->dispatchBrowserEvent('close-modal');
+                $this->dispatchBrowserEvent('alert', ['type' => 'success',  'message' => 'Sample Information deleted successfully!']);
+            }else{
+                $this->dispatchBrowserEvent('alert', ['type' => 'error',  'message' => 'Sample Information can not be deleted!']);
+            }
+
         } catch(Exception $error) {
             $this->dispatchBrowserEvent('alert', ['type' => 'error',  'message' => 'Sample Information can not be deleted!']);
         }
+        // try {
+        //     $sample = Sample::where('id', $this->delete_id)->first();
+        //     $sample->delete();
+        //     $this->delete_id = '';
+        //     $this->dispatchBrowserEvent('close-modal');
+        //     $this->dispatchBrowserEvent('alert', ['type' => 'success',  'message' => 'Sample Information deleted successfully!']);
+        // } catch(Exception $error) {
+        //     $this->dispatchBrowserEvent('alert', ['type' => 'error',  'message' => 'Sample Information can not be deleted!']);
+        // }
     }
 
     public function refresh()
